improve logger

+++ Logger.cs
@@ -5,7 +5,9 @@ namespace BetterReminders
 {
 	/// <summary>
 	/// A trivial class for diagnostic logging to a text file in usuer=profiledir/AppData/Local/OutlookBetterReminders; 
-	/// also writes to the standard Diagnostics.Trace logger as a backup, though that's not so easy to use. 
+	/// also writes to the standard Diagnostics.Trace logger as a backup, though that's not so easy to use from an addin. 
+	/// 
+	/// Since most people won't want a log, we only generate one if the file already exists
 	/// 
 	/// Currently the log level is not configurable, but as we wipe the log on startup we should never generate enough output to be a problem. 
 	/// </summary>
@@ -20,7 +22,12 @@ namespace BetterReminders
 			try
 			{
 				string logpath = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData) + @"\OutlookBetterReminders\better-reminders.log";
-				Directory.GetParent(logpath).Create();
+				if (!File.Exists(logpath))
+				{
+					// we still have diagnotics.trace logging
+					Info("Not writing a .log file as " + logpath + " does not exist");
+					return;
+				}
 				// create a new text file, overwrite existing
 				sw = new StreamWriter(new FileStream(logpath, FileMode.Create, FileAccess.Write, FileShare.ReadWrite))
 				{
